Allow locally defined (notebook) UDFs to be used as tools
